+++ api/database/seeders/UsersTableSeeder.php
@@ -11,18 +11,18 @@ class UsersTableSeeder extends Seeder
     public function run(): void
     {
         $users = [
-            ['name' => 'Super Admin', 'email' => 'superadmin@efsc-ya.test', 'password' => 'S.Admin.123'],
-            ['name' => 'Admin', 'email' => 'admin@efsc-ya.test', 'password' => 'Admin.123'],
-            ['name' => 'Developer', 'email' => 'developer@efsc-ya.test', 'password' => 'Developer.123'],
-            ['name' => 'Principal (test)', 'email' => 'principal@efsc-ya.test', 'password' => 'Test.123'],
-            ['name' => 'Vice Principal (test)', 'email' => 'viceprincipal@efsc-ya.test', 'password' => 'Test.123'],
-            ['name' => 'Section Head (test)', 'email' => 'sectionhead@efsc-ya.test', 'password' => 'Test.123'],
-            ['name' => 'Class Incharge (test)', 'email' => 'incharge@efsc-ya.test', 'password' => 'Test.123'],
-            ['name' => 'Teacher (test)', 'email' => 'teacher@efsc-ya.test', 'password' => 'Test.123'],
-            ['name' => 'Accountant (test)', 'email' => 'accountant@efsc-ya.test', 'password' => 'Test.123'],
-            ['name' => 'Computer Operator (test)', 'email' => 'operator@efsc-ya.test', 'password' => 'Test.123'],
-            ['name' => 'Parent (test)', 'email' => 'parent@efsc-ya.test', 'password' => 'Test.123'],
-            ['name' => 'Student (test)', 'email' => 'student@efsc-ya.test', 'password' => 'Test.123'],
+            ['name' => 'Super Admin', 'email' => 'superadmin@efsc-ya.com', 'password' => 'S.Admin.123'],
+            ['name' => 'Admin', 'email' => 'admin@efsc-ya.com', 'password' => 'Admin.123'],
+            ['name' => 'Developer', 'email' => 'developer@efsc-ya.com', 'password' => 'Developer.123'],
+            ['name' => 'Principal (test)', 'email' => 'principal@efsc-ya.com', 'password' => 'Test.123'],
+            ['name' => 'Vice Principal (test)', 'email' => 'viceprincipal@efsc-ya.com', 'password' => 'Test.123'],
+            ['name' => 'Section Head (test)', 'email' => 'sectionhead@efsc-ya.com', 'password' => 'Test.123'],
+            ['name' => 'Class Incharge (test)', 'email' => 'incharge@efsc-ya.com', 'password' => 'Test.123'],
+            ['name' => 'Teacher (test)', 'email' => 'teacher@efsc-ya.com', 'password' => 'Test.123'],
+            ['name' => 'Accountant (test)', 'email' => 'accountant@efsc-ya.com', 'password' => 'Test.123'],
+            ['name' => 'Computer Operator (test)', 'email' => 'operator@efsc-ya.com', 'password' => 'Test.123'],
+            ['name' => 'Parent (test)', 'email' => 'parent@efsc-ya.com', 'password' => 'Test.123'],
+            ['name' => 'Student (test)', 'email' => 'student@efsc-ya.com', 'password' => 'Test.123'],
         ];
 
         foreach ($users as $u) {